refactor(visualizator-fe): Extract lightenColor function to utils
